validaciones en Registrar y actualzar

+++ FormulariosCompras/frmActualizar.cs
@@ -145,6 +145,8 @@ namespace TRAMADE
 
         private void btnQuitar_Click(object sender, EventArgs e)
         {
+            if (!clsValidacionesCompras.validarListBox(lstProductos)) return;
+
             ObjOp.eliminarProducto(lstProductos);
             txtTotal.Text = ObjOp.TotalLista().ToString("0.00");
 
@@ -152,11 +154,8 @@ namespace TRAMADE
 
         private void btnAgregar_Click(object sender, EventArgs e)
         {
-            if (cmbProducto.SelectedIndex == -1)
-            {
-                MessageBox.Show("Seleccione un producto primero", "Aviso", MessageBoxButtons.OK, MessageBoxIcon.Warning);
-                return;
-            }
+            if (!clsValidacionesCompras.validarComboProducto(cmbProducto)) return;
+
             DataRowView drv = (DataRowView)cmbProducto.SelectedItem;
 
             ObjOp.setProducto(Convert.ToInt32(cmbProducto.SelectedValue));
@@ -208,16 +207,25 @@ namespace TRAMADE
             txtTotal.Clear();
             ObjOp.limpiarLista();
 
+            dtEntrega.Enabled = false;
+            cmbProveedor.Enabled = false;
+            cmbProducto.Enabled = false;
+            cmbFormaPago.Enabled = false;
+            nudCantidad.Enabled = false;
+
+            compraIdSeleccionado = 0;
+
+
+
         }
 
         private void btnActualizar_Click(object sender, EventArgs e)
         {
 
-           if (compraIdSeleccionado == 0)
-            {
-                MessageBox.Show("Ingrese una solicitud de compra primero","ERROR",MessageBoxButtons.OK, MessageBoxIcon.Error);
-                return;
-            }
+            //Validacines 
+            if (!clsValidacionesCompras.validarIdSeleccionado(compraIdSeleccionado)) return;
+            if (!clsValidacionesCompras.validarListBox(lstProductos)) return;
+            if (!clsValidacionesCompras.validarFechaEntrega(dtEntrega.Value)) return;
 
             try
             {
